--- api/ir/section_challenge.go
@@ -1,9 +1,0 @@
-package ir
-
-import (
-	"github.com/shu22203/internet_ranking/challenge"
-)
-
-type SectionChallenge interface {
-	AwardPointsToSubmitters() challenge.ChallengeResult
-}
